refactor: update controller updateAvatar

+++ controllers/auth-controller.js
@@ -1,20 +1,19 @@
 import bcrypt from "bcryptjs"; // хешування/верифікація пароля
 import jwt from "jsonwebtoken"; // шифрування/розшифровування токена
 import gravatar from "gravatar"; // створення тимчасової аватарки
-import path from "path";
+import path from "path"; // створення шляхів та їх нормалізація (для різних операційних систем)
 import fs from "fs/promises"; // робота з файлами
-
-import Jimp from "jimp";
 
 import { User } from "../models/user.js";
 
 import { ctrlWrapper } from "../decorators/index.js";
 
-import { HttpError } from "../helpers/index.js";
+import { HttpError, processingImage } from "../helpers/index.js";
 
 const { JWT_SECRET } = process.env;
 
 const avatarsPath = path.resolve("public", "avatars"); // абсолютний шлях до папки з файлами avatars
+const imagesExtensionsArray = ["jpg", "jpeg", "png", "bmp", "tiff", "gif"]; // підтримувані Jimp розширення зображень
 
 const register = async (req, res) => {
   const { email, password } = req.body;
@@ -108,25 +107,31 @@ const updateAvatar = async (req, res) => {
   // });
 
   // 2 варіант - з обробкою зображення за допомогою бібліотеки Jimp
-  try {
-    const uniqueFileName = `${_id}_${originalname}`; // створення унікального ім'я (з _id перед originalname)
-    const newPath = path.join(avatarsPath, uniqueFileName); // новий абсолютний шлях (з ім'ям та розширенням), де буде зберігатися файл
-
-    // Обробка зображення за допомогою бібліотеки Jimp
-    const avatar = await Jimp.read(tempPath);
-    avatar.resize(250, 250); // зміна розміру зображення на 250x250 пікселів
-    await avatar.writeAsync(newPath); // збереження після обробки файлу в папку "public/avatars"
-
-    await fs.unlink(tempPath); // видалення файлу з папки "temp"
-    const avatarURL = path.join("avatars", uniqueFileName); // створення відносного шляху (відносно корня проекту без урахування "public") для запису в модель user в БД
-    await User.findByIdAndUpdate(_id, { avatarURL });
-    res.json({
-      avatarURL,
-    });
-  } catch (error) {
-    console.error(error);
-    throw HttpError(500, "Failed to process the image");
+  // Перевірка, чи розширення зображення підтримується бібліотекою Jimp
+  const lastDotIndex = originalname.lastIndexOf(".");
+  const fileExtension = originalname.slice(lastDotIndex + 1);
+  const isImageSupport = imagesExtensionsArray.includes(fileExtension);
+  if (!isImageSupport) {
+    throw HttpError(
+      400,
+      `Sorry, this application doesn't support files with <${fileExtension}> extansion! Supported types: ${imagesExtensionsArray.join(
+        " "
+      )}.`
+    );
   }
+
+  const uniqueFileName = `${_id}_${originalname}`; // створення унікального ім'я (з _id перед originalname)
+  const newPath = path.join(avatarsPath, uniqueFileName); // новий абсолютний шлях (з ім'ям та розширенням), де буде зберігатися файл
+
+  // Обробка зображення за допомогою бібліотеки Jimp
+  await processingImage(tempPath, newPath); // зчитує і зберігає після обробки згідно вказаних шляхів
+
+  await fs.unlink(tempPath); // видалення файлу з папки "temp"
+  const avatarURL = path.join("avatars", uniqueFileName); // створення відносного шляху (відносно корня проекту без урахування "public") для запису в модель user в БД
+  await User.findByIdAndUpdate(_id, { avatarURL });
+  res.json({
+    avatarURL,
+  });
 };
 
 export default {
